requirements.txt
Django==2.2
django-mixins==0.0.10
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
mixins==0.1.1
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
